% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bear-lake.R
\docType{data}
\name{bear_lake}
\alias{bear_lake}
\title{Bear Lake Area}
\format{
\subsection{\code{bear_lake}}{

A simple feature column or \code{sfc} consisting of a single POLYGON geometry.
}
}
\source{
Coordinates digitized manually.
}
\usage{
bear_lake
}
\description{
An arbitrary bounding box drawn around Bear Lake along the border between
Utah and Idaho. See
\href{https://maps.app.goo.gl/bDBFyFhNgYbfvwuw8}{location on Google Maps}.
}
\keyword{datasets}
