% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\alias{element_codes}
\alias{network_codes}
\title{Element Codes}
\format{
\subsection{\code{element_codes}}{

A \code{tibble} with columns \code{code}, \code{name}, and \code{unit}.
}

\subsection{\code{network_codes}}{

A \code{tibble} with columns \code{code} and \code{description}.
}
}
\source{
\href{https://www.nrcs.usda.gov/sites/default/files/2023-03/AWDB\%20Web\%20Service\%20User\%20Guide.pdf}{AWDB Web Service User Guide}
}
\usage{
element_codes

network_codes
}
\description{
Codes for station elements and networks.
}
\keyword{datasets}
