% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{bear_lake}
\alias{cascades}
\title{Datasets}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\source{
Coordinates digitized manually.
}
\usage{
bear_lake

cascades
}
\description{
Arbitrary bounding boxes drawn around potential areas of interest.
}
\details{
Areas of interest include:
\itemize{
\item Bear Lake along the border between Utah and Idaho. See
\href{https://maps.app.goo.gl/bDBFyFhNgYbfvwuw8}{location on Google Maps}.
\item Cascades in west central Oregon. See
\href{https://maps.app.goo.gl/sBtTHaN8SeM6xJWv7}{location on Google Maps}.
}
}
\keyword{datasets}
