% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-station-data.R
\name{get_station_data}
\alias{get_station_data}
\title{Get Station Data From USDA NWCC AWDB}
\usage{
get_station_data(
  aoi,
  elements,
  networks = "*",
  duration = "DAILY",
  begin_date,
  end_date,
  request_size = 10L,
  as_sf = FALSE
)
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{networks}{character vector, abbreviations or codes for station networks
of interest (e.g., "USGS" refers to all USGS soil monitoring stations).
Default is \code{*}, for "all networks". See Details for available networks and
codes.}

\item{duration}{character scalar, the temporal resolution of the element
measurements. Available values include \code{DAILY} (default), \code{HOURLY},
\code{SEMIMONTHLY}, \code{MONTHLY}, \code{CALENDAR_YEAR}, \code{WATER_YEAR}.}

\item{begin_date}{character scalar, start date for time period of interest.
Date must be in format \code{"YYYY-MM-DD"}.}

\item{end_date}{character scalar, end date for time period of interest. Date
must be in format \code{"YYYY-MM-DD"}.}

\item{request_size}{integer scalar, number of individual stations to include
in each query. This helps to meet rate limits imposed by the API. If you are
getting a request error, you might try lowering this number. Default is
\code{10L}.}

\item{as_sf}{boolean scalar, whether to return the data as an \code{sf} table.
Default is \code{FALSE}. Repeating the spatial data across each station element
and its time series can be costly.}
}
\value{
if \code{as_sf}, an \code{sf} table, otherwise a simple data.frame. The number
of rows depends on the number of stations and element parameters. Time series
data are included as a list column named \code{"values"}.
}
\description{
Get station data from the USDA National Water and Climate Center Air and
Water Database REST API.
}
\details{
TODO: add table of possible networks and elements
}
\examples{
\dontrun{
library(sf)

pnw <- sf::st_bbox(c(
  xmin = -123.9745334,
  ymin = 42.689170,
  xmax = -117.212106,
  ymax = 48.591128
))

pnw <- sf::st_as_sfc(pnw, crs = 4326)

get_station_data(pnw, elements = "WTEQ", networks = "SNTL")
}
}
