% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-station-metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Get Station Metadata From USDA NWCC AWDB}
\usage{
get_station_metadata(
  aoi,
  elements,
  networks = "*",
  durations = "DAILY",
  return_forecast_metadata = FALSE,
  return_reservoir_metadata = FALSE,
  return_element_metadata = FALSE,
  request_size = 10L
)
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{networks}{character vector, abbreviations or codes for station networks
of interest (e.g., "USGS" refers to all USGS soil monitoring stations).
Default is \code{*}, for "all networks". See Details for available networks and
codes.}

\item{return_forecast_metadata}{scalar boolean, whether to return a list
column with forecast metadata. Default is \code{FALSE}.}

\item{return_reservoir_metadata}{scalar boolean, whether to return a list
column with reservoir metadata. Default is \code{FALSE}.}

\item{return_element_metadata}{scalar boolean, whether to return a list
column with element metadata. Default is \code{FALSE}.}

\item{request_size}{integer scalar, number of individual stations to include
in each query. This helps to meet rate limits imposed by the API. If you are
getting a request error, you might try lowering this number. Default is
\code{10L}.}
}
\value{
an \code{sf} table with station data. The number of rows is equal to the
number of stations found. If requested, forecast, reservoir,
and element metadata for each station are stored in their respective list
columns.
}
\description{
Get station metadata from the USDA National Water and Climate Center Air and
Water Database REST API. Each station is uniquely identified by a station
triplet of the form \code{station:state:network}.
}
\details{
If you just want the locations, set all metadata options to \code{FALSE}.
}
\examples{
\dontrun{
library(sf)

pnw <- sf::st_bbox(c(
  xmin = -123.9745334,
  ymin = 42.689170,
  xmax = -117.212106,
  ymax = 48.591128
))

pnw <- sf::st_as_sfc(pnw, crs = 4326)

# get locations and element metadata for stations with snow water equivalent
# (WTEQ) and soil moisture percent (SMS:*, * = any depth) measurements
get_station_metadata(
  pnw,
  elements = c("WTEQ", "SMS:*"),
  return_element_metadata = TRUE
)

# get just locations for stations with snow water equivalent measurements
get_station_metadata(pnw, elements = "WTEQ")
}
}
