% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awdb.R
\docType{package}
\name{awdb}
\alias{awdb-package}
\alias{awdb}
\title{Query the USDA AWDB REST API}
\description{
The awdb package implements functions to interface with the AWDB REST API
from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kbvernon.github.io/awdb/}
}

}
\author{
\strong{Maintainer}: Kenneth Blake Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID})

}
\keyword{internal}
