% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecasts.R
\name{get_forecasts}
\alias{get_forecasts}
\title{Get Station Forecasts From USDA NWCC AWDB}
\usage{
get_forecasts(aoi, elements, awdb_options = set_options(), as_sf = FALSE)
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{awdb_options}{an \code{awdb_options} list with additional query parameters.}

\item{as_sf}{boolean scalar, whether to return the data as an \code{sf} table.
Default is \code{FALSE}. Repeating the spatial data across each station element
and its time series can be costly.}
}
\value{
if \code{as_sf}, an \code{sf} table, otherwise a simple data.frame. The number
of rows depends on the number of stations and element parameters. Time series
data are included as a list column named \code{"forecast_values"}.
}
\description{
Get station forecasts from the USDA National Water and Climate Center Air and
Water Database REST API.
}
\details{
TODO!
}
\examples{
# get forecasts for snow water equivalent (WTEQ)
get_forecasts(bear_lake, elements = "WTEQ")

}
