% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-forecasts.R
\name{get_forecasts}
\alias{get_forecasts}
\title{Get Station Forecasts}
\usage{
get_forecasts(
  aoi = NULL,
  elements,
  awdb_options = set_options(),
  as_sf = FALSE
)
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}. If \code{NULL} (the default),
no spatial filter is performed.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{awdb_options}{an \code{awdb_options} list with additional query parameters.}

\item{as_sf}{boolean scalar, whether to return the data as an \code{sf} table.
Default is \code{FALSE}. Repeating the spatial data across each station element
and its time series can be costly.}
}
\value{
if \code{as_sf}, an \code{sf} table, otherwise a simple data.frame. The number
of rows depends on the number of stations and element parameters. Time series
data are included as a list column named \code{"forecast_values"}.
}
\description{
Get station forecasts from the USDA National Water and Climate Center Air and
Water Database REST API. These will almost always be streamflow forecasts,
set with \code{elements = "SRVO"}, but some others are also available, albeit with
extremely limited spatial representation (see Details).
}
\details{
This endpoint will accept the following query parameters via \code{set_options()}:
\itemize{
\item \code{begin_publication_date}
\item \code{end_publication_date}
\item \code{exceedence_probabilities}
\item \code{forecast_periods}
}

The following can also be passed to filter stations:
\itemize{
\item \code{station_names}
\item \code{dco_codes}
\item \code{county_names}
\item \code{hucs}
\item \code{active_only}
}

You may also specify \code{networks} and \code{request_size}. The \code{networks} parameter
is used internally to build unique station triplet identifiers of the form
\code{station:state:network} which are then passed to the endpoint, so it serves
to filter stations to just those networks. The \code{request_size} parameter is
for handling rate limits, which are based on the number of elements - a hard
value to measure directly, so this parameter is more a rule of thumb than a
strict standard. If processing is slow for you, you may find experimenting
with this parameter useful.

Note that the \code{duration} parameter is ignored - or, more precisely, it is set
to \code{NULL}.

See \code{set_options()} for more details.
\subsection{Forecast Elements}{

Almost all forecasts are reported in \verb{SRVO``, the adjusted streamflow set which accounts for upstream operations such as reservoir operations and diversions. }JDAY\verb{, }RESC\verb{, and }REST\verb{are hardly used at all, mostly to maintain historical forecasts made at Lake Tahoe (the birthplace of the snow survey). In general, it's recommended to use}SRVO`.
}
}
\examples{
# get streamflow forecasts
get_forecasts(cascades, elements = "SRVO")

}
