% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-station-data.R
\name{get_station_data}
\alias{get_station_data}
\title{Get Station Data From USDA NWCC AWDB}
\usage{
get_station_data(station_triplets, elements, begin_date, end_date)
}
\arguments{
\item{station_triplets}{character vector, a vector of station triplets having
the form \code{station:state:network}, with \code{state} being the state code for each
station (e.g., "OR" for "Oregon") and \code{network} the abbreviation of the
station network that each station is part of (e.g., "USGS" refers to all USGS
soil monitoring stations). Note that the \code{*} wildcard can be used to select
all of \code{station}, \code{state}, or \code{network}. The triplet \verb{*:*:*} will return the
entire database. See Details for all networks.}

\item{elements}{character vector, the abbreviation of the variable for which
measurements are sought (e.g., "SMS" for "Soil Moisture Percent"). See
Details for all elements.}
}
\value{
a `data.frame`` with station data. The number of rows depends on the
number of stations and elements measured at each. Time series data are
included as a list column named "values" (the name used by the API).
}
\description{
Get station data from the USDA National Water and Climate Center Air and
Water Database REST API. Each station is uniquely identified by a station
triplet of the form \code{station:state:network}.
}
\details{
TODO: add table of possible networks
}
\examples{
get_station_data("302:OR:SNTL", elements = "WTEQ")

}
