% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-station-metadata.R
\name{get_station_metadata}
\alias{get_station_metadata}
\title{Get Station Metadata From USDA NWCC AWDB}
\usage{
get_station_metadata(station_triplets)
}
\arguments{
\item{station_triplets}{character vector, a vector of station triplets having
the form \code{station:state:network}, with \code{state} being the state code for each
station (e.g., "OR" for "Oregon") and \code{network} the abbreviation of the
station network that each station is part of (e.g., "SCAN" for "Soil Climate
Adaptation Network"). Note that the \code{*} wildcard can be used to select all of
\code{station}, \code{state}, or \code{network}. The triplet \verb{*:*:*} will return the entire
database. See Details for all networks.}
}
\value{
a \code{data.frame} with number of rows equal to
\code{length(station_triplets)}. Element metadata for each station is stored in a
list column named "station_elements".
}
\description{
Get station metadata from the USDA National Water and Climate Center Air and
Water Database REST API. Each station is uniquely identified by a station
triplet of the form \code{station:state:network}.
}
\examples{
get_station_metadata("302:OR:SNTL")

}
