% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stations.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations From USDA NWCC AWDB in Area of Interest}
\usage{
get_stations(aoi, elements, awdb_options = set_options())
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{awdb_options}{an \code{awdb_options} list with additional query parameters.}
}
\value{
an \code{sf} table with station data.
}
\description{
Get stations and their coordinates from the Air and Water Database REST API
that fall in area of interest.
}
\details{
You can also subset stations using query parameters in \code{\link[=set_options]{set_options()}}.
}
\examples{
# get stations with WTEQ elements
get_stations(
  bear_lake,
  elements = "WTEQ"
)

# get stations with WTEQ elements that are in the SNTL (and SCAN) network
get_stations(
  bear_lake,
  elements = "WTEQ",
  awdb_options = set_options(networks = "SNTL")
)

# get stations with WTEQ elements that are measured daily
get_stations(
  bear_lake,
  elements = "WTEQ",
  awdb_options = set_options(durations = "DAILY")
)

}
