% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stations.R
\name{get_stations}
\alias{get_stations}
\title{Get Station Metadata}
\usage{
get_stations(aoi, elements, awdb_options = set_options())
}
\arguments{
\item{aoi}{\code{sfc} POLYGON scalar, the area of interest used for performing
a spatial filter on available stations in \code{network}.}

\item{elements}{character vector, abbreviations or codes for variables of
interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available
elements and codes.}

\item{awdb_options}{an \code{awdb_options} list with additional query parameters.}
}
\value{
an \code{sf} table with station metadata.
}
\description{
Get station metadata from the USDA National Water and Climate Center Air and
Water Database REST API. This includes their spatial coordinates.
}
\details{
This endpoint will accept the following query parameters via \code{set_options()}:
\itemize{
\item \code{station_names}
\item \code{dco_codes}
\item \code{county_names}
\item \code{hucs}
\item \code{return_forecast_metadata}
\item \code{return_reservoir_metadata}
\item \code{return_element_metadata}
\item \code{active_only}
}

You may also specify \code{networks}. The \code{networks} parameter is used internally
to build unique station triplet identifiers of the form
\code{station:state:network}, so it serves to filter stations to just those
networks.

See \code{set_options()} for more details.
}
\examples{
# get stations with WTEQ elements
get_stations(
  bear_lake,
  elements = "WTEQ"
)

# get stations with WTEQ elements that are in the SNTL network
get_stations(
  bear_lake,
  elements = "WTEQ",
  awdb_options = set_options(networks = "SNTL")
)

}
